v0.7.0: device identity, rdma-core device attributes, MRRS control
- ibv_query_device fix: vendor_part_id / fw_ver / device id / revision are now
  real (fw 14.22.1002, part 0x1015); fwRev widened to 64 bits and re-read after FLR
- ibv_device_attr now follows rdma-core 50.0 (ABI change): fw_ver char[64],
  64-bit max_mr_size, device_cap_flags, atomic_cap, max_mw/max_ah/max_srq_wr/
  max_srq_sge/max_pkeys, local_ca_ack_delay
- New diagnostic UserClient method kMlxUCMethodPcieDevCtl (0x10be): PCIe
  Device Control readback + Max Read Request Size override, re-applied after FLR
- New tools mlx_mrrs (MRRS read/set/restore) and mlx_lat (two-machine latency)
- Docs: CHANGELOG (MCDMA / mcdma-dsv41 results, MRRS findings) and refreshed
  inference-client-guide; DEXT 0.550

+++ src/dext/Sources/hw/MlxHCA.hpp
@@ -36,7 +36,8 @@ struct MlxIBAddr {
  * Hardware capability summary — key fields returned by QUERY_HCA_CAP
  */
 struct MlxHcaCaps {
-    uint32_t fwRev;
+    /* major << 32 | minor << 16 | subminor, as mlx5_ib reports fw_ver */
+    uint64_t fwRev;
     uint16_t cmdifRev;
     uint8_t  portType;
     uint32_t numPorts;
